Uploading code and executable
Load an image with objects you would like to count. Use the mouse to click the image and counts the number of clicks.

I would like to find a way to add a marking on top of the image/button to make counting easier, but it is still better than counting in your head